DUOS-977 moved utility functions out of resource and into InstitutionUtil, refactored resource and tests
