+++ src/main/java/org/broadinstitute/consent/http/resources/InstitutionResource.java
@@ -1,9 +1,6 @@
 package org.broadinstitute.consent.http.resources;
 
-import com.google.gson.ExclusionStrategy;
-import com.google.gson.FieldAttributes;
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.google.inject.Inject;
 import io.dropwizard.auth.Auth;
 import java.util.List;
@@ -18,23 +15,24 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
-import org.broadinstitute.consent.http.enumeration.UserRoles;
 import org.broadinstitute.consent.http.models.AuthUser;
 import org.broadinstitute.consent.http.models.Institution;
 import org.broadinstitute.consent.http.models.User;
-import org.broadinstitute.consent.http.models.UserRole;
 import org.broadinstitute.consent.http.service.InstitutionService;
 import org.broadinstitute.consent.http.service.UserService;
+import org.broadinstitute.consent.http.util.InstitutionUtil;
 
 @Path("api/institutions")
 public class InstitutionResource extends Resource {
   private final UserService userService;
   private final InstitutionService institutionService;
+  private final InstitutionUtil institutionUtil;
 
   @Inject
-  public InstitutionResource(UserService userService, InstitutionService institutionService) {
+  public InstitutionResource(UserService userService, InstitutionService institutionService, InstitutionUtil institutionUtil) {
     this.userService = userService;
     this.institutionService = institutionService;
+    this.institutionUtil = institutionUtil;
   }
 
   @GET
@@ -43,8 +41,8 @@ public class InstitutionResource extends Resource {
   public Response getInstitutions(@Auth AuthUser authUser) {
     try{
       User user = userService.findUserByEmail(authUser.getName());
-      Boolean isAdmin = checkIfAdmin(user);
-      Gson gson = getGsonBuilder(isAdmin);
+      Boolean isAdmin = institutionUtil.checkIfAdmin(user);
+      Gson gson = institutionUtil.getGsonBuilder(isAdmin);
       List<Institution> institutions = institutionService.findAllInstitutions();
       return Response.ok().entity(gson.toJson(institutions)).build();
     } catch(Exception e) {
@@ -59,8 +57,8 @@ public class InstitutionResource extends Resource {
   public Response getInstitution(@Auth AuthUser authUser, @PathParam("id") String paramId) {
     try{
       User user = userService.findUserByEmail(authUser.getName());
-      Boolean isAdmin = checkIfAdmin(user);
-      Gson gson = this.getGsonBuilder(isAdmin);
+      Boolean isAdmin = institutionUtil.checkIfAdmin(user);
+      Gson gson = institutionUtil.getGsonBuilder(isAdmin);
       Integer id = Integer.parseInt(paramId);
       Institution institution = institutionService.findInstitutionById(id);
       return Response.ok().entity(gson.toJson(institution)).build();
@@ -115,39 +113,5 @@ public class InstitutionResource extends Resource {
     } catch(Exception e) {
       return createExceptionResponse(e);
     }
-  }
-
-  // Gson builder and exclusion strategy helpers
-  // Opting to not null values, null has the implication of an absence of value
-  // Whereas the absence of a field can mean an absence of value OR an omission of data
-  private Gson getGsonBuilder(Boolean isAdmin) {
-    ExclusionStrategy strategy = getSerializationExclusionStrategy(isAdmin);
-    return new GsonBuilder().addSerializationExclusionStrategy(strategy).create();
-  }
-
-  private ExclusionStrategy getSerializationExclusionStrategy(Boolean isAdmin) {
-    return new ExclusionStrategy() {
-      @Override
-      public boolean shouldSkipField(FieldAttributes field) {
-        String fieldName = field.getName();
-        if (!isAdmin && (fieldName != "id" && fieldName != "name")) {
-          return true;
-        } else {
-          return false;
-        }
-      }
-
-      //NOTE: shouldSkipClass is mandatory when creating an ExclusionStrategy
-      //No reason to skip class (only dealing with Institution), so you can just return false here
-      @Override
-      public boolean shouldSkipClass(Class<?> c) {
-        return false;
-      }
-    };
-  }
-
-  private Boolean checkIfAdmin(User user) {
-    List<UserRole> roles = user.getRoles();
-    return roles.stream().anyMatch((userRole) -> userRole.getRoleId() == UserRoles.ADMIN.getRoleId());
-  }
+  };
 }
